Fix problem with 'pools' command
There was a problem with a mismatch between the command and the
dCache message.  The command operates on a single poolgroup but
the dCache message accepts multiple poolgroups.

The solution is to create a Set object containing the pool requested
by the user.  Set is a subclass of Iterable, so satifying the
message's constructor.